notification 구현 및 소소한 수정

+++ app/src/main/java/com/kang/floapp/view/main/frag/storage/FragStorage.java
@@ -101,8 +101,13 @@ public class FragStorage extends Fragment {
             @Override
             public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int direction) {
                 Log.d(TAG, "onSwiped: " + viewHolder.getAdapterPosition());
+
                 int id = storageAdapter.getStorageId(viewHolder.getAdapterPosition());
-                mainViewModel.deleteByStorageId(id);
+
+                if(id != 0) {
+                    mainViewModel.deleteByStorageId(id);
+                }
+
             }
         };
